+++ database/seeders/DatabaseSeeder.php
@@ -12,5 +12,6 @@ final class DatabaseSeeder extends Seeder
         $this->call(SaasSeeder::class);
         $this->call(CmsSeeder::class);
         $this->call(MemberSettingsSeeder::class);
+        $this->call(AiSeeder::class);
     }
 }